fix: 🐛 행사 목록 skeleton으로 변경
