+++ src/components/common/BottomSpace.tsx
@@ -1,0 +1,3 @@
+export default function BottomSpace() {
+  return <div className="h-14"></div>;
+}
