+++ src/components/events/EventCardSkeleton.tsx
@@ -1,0 +1,3 @@
+export default function EventCardSkeleton() {
+  return <div className="min-h-72 animate-pulse rounded-3xl bg-gray-200 dark:bg-gray-700"></div>;
+}
